+++ src/components/FeedbackBox/feedbackBoxChangelogData.ts
@@ -10,6 +10,18 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Styles"],
+    notes: [
+      "Updates to suppress native browser validation.",
+      "Utilizies onChange events to clear error state in comment and email input elements.",
+      "Updates component and documentation example to use sentence case.",
+      "Updates the email-specific confirmation message.",
+    ],
+  },
+  {
     date: "2025-08-11",
     version: "4.0.0",
     type: "Update",