+++ src/components/Icons/iconChangelogData.ts
@@ -10,6 +10,16 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Accessibility", "Functionality"],
+    notes: [
+      "Added the `navigationArrowDown`, `contentBookmarkOutline`, and `contentBookmark` icons.",
+      "Updated the following icon options to remove the ID attributes: `actionHome`, `actionInfo`, `actionList`, `actionPayment`, `actionSearch`, `contentCopy`, `navigationApps`, `navigationMoreVert`, `navigationSubdirectoryArrowLeft`, and `navigationSubdirectoryArrowRight`.",
+    ],
+  },
+  {
     date: "2026-05-20",
     version: "4.1.7",
     type: "Update",