+++ src/components/HelperErrorText/helperErrorTextChangelogData.ts
@@ -10,6 +10,15 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "2026-06-09",
+    version: "4.2.0",
+    type: "Update",
+    affects: ["Accessibility", "Documentation", "Functionality"],
+    notes: [
+      "Removed the default `aria-atomic` and `aria-live` attributes from the helper text div.",
+    ],
+  },
+  {
     date: "2025-08-11",
     version: "4.0.0",
     type: "Update",