+++ CHANGELOG.md
@@ -10,7 +10,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Adds
 
-- Adds the `navigationArrowDown`, `contentBookmarkOutline`, and `contentBookmark` options to the `Icon` component.
+- Adds the `contentBookmark`, `contentBookmarkOutline`, `mapsLayers`, and `navigationArrowDown` options to the `Icon` component.
+- Adds `Placeholder Images` to the `Style Guide`.
 
 ### Updates